% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlc.R
\name{dlc}
\alias{dlc}
\title{Compute D, L, and C in Fisher Optimal Partitions Algorithm}
\usage{
dlc(X, maxk)
}
\arguments{
\item{X}{a set of samples}

\item{maxk}{maximum number of k}
}
\value{
The D, L, and C statistics in
Fisher optimal partitions algorithm
}
\description{
Compute D, L, and C in Fisher Optimal Partitions Algorithm
}

