% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHPL-package.R
\docType{package}
\name{OHPL-package}
\alias{OHPL-package}
\title{Ordered Homogeneity Pursuit Lasso (OHPL)}
\description{
\tabular{ll}{Package: \tab OHPL\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}

