% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohpl.R
\name{ohpl}
\alias{ohpl}
\title{Ordered Homogeneity Pursuit Lasso (OHPL)}
\usage{
ohpl(X.cal, y.cal, maxcomp, gamma, X.test, y.test, cv.folds = 5L, G = 30L,
  type = c("max", "median"), scale = TRUE, pls.method = "simpls")
}
\arguments{
\item{X.cal}{Predictor matrix (training)}

\item{y.cal}{Response matrix with one column (training)}

\item{maxcomp}{Maximum number of components for PLS}

\item{gamma}{A number between (0, 1) for generating
the gamma sequence. An usual choice for gamma could be
\code{n * 0.05}, where \code{n} is a number in 2, 3, ..., 19.}

\item{X.test}{X.test Predictor matrix (test)}

\item{y.test}{y.test Response matrix with one column (test)}

\item{cv.folds}{Number of cross-validation folds}

\item{G}{Maximum number of variable groups}

\item{type}{Find the maximum absolute correlation (\code{"max"})
or find the median of absolute correlation (\code{"median"}).
Default is \code{"max"}.}

\item{scale}{Should the predictor matrix be scaled?
Default is \code{TRUE}.}

\item{pls.method}{Method for fitting the PLS model.
Default is \code{"simpls"}. See the details section
in \code{\link[pls]{plsr}} for all possible options.}
}
\value{
model list
}
\description{
This function fits the ordered homogeneity pursuit lasso (OHPL) model.
It streamlines parameter tuning with cross-validation, model fitting,
and making predictions on the new data.
}
\examples{
data("wheat")

X = wheat$x
y = wheat$protein
n = nrow(wheat$x)

set.seed(1001)
samp.idx = sample(1L:n, round(n * 0.7))
X.cal = X[samp.idx, ]
y.cal = y[samp.idx]
X.test = X[-samp.idx, ]
y.test = y[-samp.idx]

# needs to run for a while
\dontrun{
fit = ohpl(
  X.cal, y.cal, maxcomp = 10, gamma = 0.8,
  X.test, y.test, G = 30, type = "max")

fit$RMSEP}
}

