% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlc.R
\name{dlc}
\alias{dlc}
\title{Compute D, L, and C in the Fisher optimal partitions algorithm}
\usage{
dlc(X, maxk)
}
\arguments{
\item{X}{A set of samples.}

\item{maxk}{Maximum number of \code{k}.}
}
\value{
The D, L, and C statistics in the
Fisher optimal partitions algorithm.
}
\description{
Compute D, L, and C in the Fisher optimal partitions algorithm
}
