% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohpl-package.R
\docType{package}
\name{ohpl-package}
\alias{ohpl-package}
\title{Ordered Homogeneity Pursuit Lasso (OHPL)}
\description{
\tabular{ll}{ Package: \tab ohpl\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}

