% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHPL-package.R
\docType{package}
\name{OHPL-package}
\alias{OHPL-package}
\title{OHPL: Ordered Homogeneity Pursuit Lasso for Group Variable Selection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Ordered homogeneity pursuit lasso (OHPL) algorithm for group variable selection proposed in Lin et al. (2017) \doi{10.1016/j.chemolab.2017.07.004}. The OHPL method exploits the homogeneity structure in high-dimensional data and enjoys the grouping effect to select groups of important variables automatically. This feature makes it particularly useful for high-dimensional datasets with strongly correlated variables, such as spectroscopic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohpl.io}
  \item \url{https://ohpl.io/doc/}
  \item \url{https://github.com/nanxstats/OHPL}
  \item Report bugs at \url{https://github.com/nanxstats/OHPL/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item You-Wu Lin \email{lyw015813@126.com}
}

}
\keyword{internal}
